+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* Fix: QGIS 3.34, type object 'Qgis' has no attribute 'FeatureRequestFlag'
+
 ## 2.15.2 - 2026-07-28
 
 * Fix: Invalid widget config for embedded layers